+++ src/position/game_state.cpp
@@ -26,32 +26,32 @@ void game_state::reset()
     root_ply = {};
 }
 
-std::span<const std::uint32_t> game_state::get_pv(std::size_t ply) noexcept
+std::span<const std::uint64_t> game_state::get_pv(std::size_t ply) noexcept
 {
     // Initialise the return value with the stored PV, possibly containing illegal moves.
-    std::span<const std::uint32_t> ret = pv.table[ply];
+    std::span<const std::uint64_t> ret = pv.table[ply];
 
     std::size_t legal_moves {};
-    std::vector<std::uint32_t> unmake_buf;
+    std::vector<std::uint64_t> unmake_buf;
     while (legal_moves < ret.size())
     {
         // Return early if we're in a repetition-draw.
         if (is_repetition_draw())
             break;
 
-        const std::uint32_t move { ret[legal_moves] };
+        const std::uint64_t move { ret[legal_moves] };
 
         // Is this move in the list of pseudo-legal moves?
-        std::array<std::uint32_t, MAX_MOVES_PER_POSITION> move_buf;
+        std::array<std::uint64_t, MAX_MOVES_PER_POSITION> move_buf;
         const std::size_t moves { generate_pseudo_legal_moves(bb, move_buf) };
-        const std::span<const std::uint32_t> move_list { move_buf.data(), moves };
+        const std::span<const std::uint64_t> move_list { move_buf.data(), moves };
 
         // Exit immediately if this move isn't even pseudo-legal.
         if (std::find(move_list.begin(), move_list.end(), move) == move_list.end())
             break;
 
         // Make sure the move is fully-legal after playing it and checking legality.
-        std::uint32_t unmake;
+        std::uint64_t unmake;
         const bool legal { make_move({ .check_legality=true }, *this, move, unmake) };
         unmake_buf.push_back(unmake);
         if (!legal)
@@ -61,7 +61,7 @@ std::span<const std::uint32_t> game_state::get_pv(std::size_t ply) noexcept
     }
 
     // Restore our game-state to how it was before the call.
-    std::for_each(unmake_buf.rbegin(), unmake_buf.rend(), [this] (const std::uint32_t& unmake) { unmake_move(*this, unmake); });
+    std::for_each(unmake_buf.rbegin(), unmake_buf.rend(), [this] (const std::uint64_t& unmake) { unmake_move(*this, unmake); });
 
     return ret.subspan(0, legal_moves);
 }
