+++ src/utility/perft.cpp
@@ -9,7 +9,7 @@
 namespace
 {
 
-std::size_t perft_recursive_unmake_no_hash(bitboard& bb, std::size_t depth, std::span<std::uint32_t> move_buf)
+std::size_t perft_recursive_unmake_no_hash(bitboard& bb, std::size_t depth, std::span<std::uint64_t> move_buf)
 {
     if (depth == 0) [[unlikely]]
         return 1;
@@ -20,7 +20,7 @@ std::size_t perft_recursive_unmake_no_hash(bitboard& bb, std::size_t depth, std:
 
     for (std::size_t i = 0; i < moves; i++)
     {
-        std::uint32_t unmake;
+        std::uint64_t unmake;
         if (make_move({ .check_legality = true }, bb, move_buf[i], unmake)) [[likely]]
             ret += perft_recursive_unmake_no_hash(bb, depth-1, move_buf.subspan(moves));
 
@@ -38,7 +38,7 @@ struct perft_value_type
 
 details::hash_table<perft_value_type> perft_hash_table;
 
-std::size_t perft_recursive_unmake_hash(bitboard& bb, std::uint64_t& hash, std::size_t depth, std::span<std::uint32_t> move_buf)
+std::size_t perft_recursive_unmake_hash(bitboard& bb, std::uint64_t& hash, std::size_t depth, std::span<std::uint64_t> move_buf)
 {
     if (depth == 0) [[unlikely]]
         return 1;
@@ -54,7 +54,7 @@ std::size_t perft_recursive_unmake_hash(bitboard& bb, std::uint64_t& hash, std::
 
     for (std::size_t i = 0; i < moves; i++)
     {
-        std::uint32_t unmake;
+        std::uint64_t unmake;
         if (make_move({ .check_legality = true }, bb, move_buf[i], unmake, hash)) [[likely]]
             ret += perft_recursive_unmake_hash(bb, hash, depth-1, move_buf.subspan(moves));
 
@@ -82,7 +82,7 @@ std::size_t get_perft_hash_table_bytes()
 
 std::size_t perft(const bitboard& start, std::size_t depth)
 {
-    std::vector<std::uint32_t> move_buf(depth*MAX_MOVES_PER_POSITION);
+    std::vector<std::uint64_t> move_buf(depth*MAX_MOVES_PER_POSITION);
 
     bitboard bb { start };
