+++ tests/test_move_generation.cpp
@@ -19,7 +19,7 @@ void test_move_generation_recursive(const bitboard& bb, const std::uint64_t hash
 
     // Our buffer of moves to loop over will also include a null-moves.
     constexpr std::size_t max_moves_in_position_including_null_move { MAX_MOVES_PER_POSITION+1 };
-    std::vector<std::uint32_t> move_buf(depth*max_moves_in_position_including_null_move);
+    std::vector<std::uint64_t> move_buf(depth*max_moves_in_position_including_null_move);
 
     // Generate our move-list and add a null-move on the end to test.
     std::size_t moves { generate_pseudo_legal_moves(bb, move_buf) };
@@ -27,9 +27,9 @@ void test_move_generation_recursive(const bitboard& bb, const std::uint64_t hash
 
     for (std::size_t i = 0; i < moves; i++)
     {
-        const std::uint32_t move = move_buf[i];
+        const std::uint64_t move = move_buf[i];
 
-        std::uint32_t unmake;
+        std::uint64_t unmake;
         make_move({ .check_legality = true }, bb_copy, move, unmake, hash_copy);
 
         test_move_generation_recursive(bb_copy, hash_copy, depth-1);
